requirements.txt
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.1.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.56.0
idna==3.10
importlib-resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
mysql-connector-python==9.0.0
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
tkcalendar==1.6.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.92
ultralytics-thop==2.0.14
urllib3==2.2.3
waitress==3.0.0
werkzeug==3.0.6
zipp==3.20.2